Clock.requestUpdate does not affect updateDuration anymore.
It's more logical to have the updateDuration as a global parameter of the
clock, and to have the requestUpdate method to be a simple trigger for the
next coming tick. This way, a clock that will update constantly for 10
seconds, can be updated for 30 seconds (because some current event requires
that) with just one call to requestUpdate.